requirements.txt
Django~=2.2.12
django-markdownify==0.8.0